requirements.txt
numpy==1.26.4
pandas==2.2.2
Pillow
ultralytics
opencv-python
matplotlib
